Testing doughnut graph on budget page, data and formatting. Line chart is not yet done, but messages are looking cute.
Not sure how I want to layout the budget page. The basic information about a budget is small, but very important. The chart is large, and also important, but I don't think they are important enough to come first on the page. As things are right now, however, it is way to vertical and distracting compared to a more horizontal layout to mirror the horizontal-ness of a screen.